Split and restructure debugger code
