+++ platforms/ESP-IDF/CMakeLists.txt
@@ -1,5 +1,10 @@
 set(SOURCE_FILES
         ../../src/Debug/debugger.cpp
+        ../../src/Debug/debugger-command.cpp
+        ../../src/Debug/debugger-snapshot.cpp
+        ../../src/Debug/debugger-legacy.cpp
+        ../../src/Debug/debugger-proxy.cpp
+        ../../src/Debug/debugger-overrides.cpp
         ../../src/Debug/nanopb_encoder.cpp
         ../../src/Debug/nanopb/debug.pb.c
         ../../src/Debug/nanopb/pb_common.c
