+++ src/Debug/debugger-private.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include "debugger.h"
+
+#include <algorithm>
+#include <cinttypes>
+#include <cstring>
+#include <optional>
+
+#include "../Memory/mem.h"
+#include "../Utils//util.h"
+#include "../Utils/macros.h"
+#include "../WARDuino/CallbackHandler.h"
+#include "nanopb_encoder.h"
